+++ back/src/main/java/com/question/test/TestQuestionRestController.java
@@ -4,6 +4,7 @@ import com.GeneralRestController;
 import com.question.test.test_answer.TestAnswer;
 import com.question.test.test_question.TestQuestion;
 import com.unit.Unit;
+import org.aspectj.weaver.ast.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -26,7 +27,7 @@ public class TestQuestionRestController extends GeneralRestController {
     }
 
     @GetMapping("/{questionID}")
-    public ResponseEntity<TestQuestion> getListQuestion(@PathVariable long unitID, @PathVariable long questionID) {
+    public ResponseEntity<TestQuestion> getTestQuestion(@PathVariable long unitID, @PathVariable long questionID) {
         Optional<Unit> unit = this.unitService.findOne(unitID);
         Optional<TestQuestion> question = this.testQuestionService.findOne(questionID);
 
@@ -107,5 +108,19 @@ public class TestQuestionRestController extends GeneralRestController {
     @GetMapping("/{id}/answer/user/{userId}")
     public ResponseEntity<List<Object>> getUserAnswers(@PathVariable long id, @PathVariable long userId) {
         return new ResponseEntity<>(this.testQuestionService.findUserAnswers(userId, id), HttpStatus.OK);
+    }
+
+
+
+    @GetMapping(value = "/{questionID}/chosenWrongAnswersCount")
+    public ResponseEntity<Object> getChosenWrongAnswersCount(@PathVariable long unitID, @PathVariable long questionID) {
+        Optional<Unit> unit = this.unitService.findOne(unitID);
+        Optional<TestQuestion> question = this.testQuestionService.findOne(questionID);
+
+        if (unit.isPresent() && question.isPresent()) {
+            return new ResponseEntity<>(this.testQuestionService.findChosenWrongAnswersCount(questionID), HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }