+++ front/src/app/app.routing.ts
@@ -23,6 +23,7 @@ import {ModuleFormComponent} from './itinerary/module/moduleForm/module-form.com
 import {UnitsBlocksToolComponent} from './itinerary/module/moduleEditor/units-blocks-tool.component';
 import {ImageComponent} from './images/image.component';
 import {LessonSlidesToolComponent} from './itinerary/lesson/lessonTools/lesson-slides-tool.component';
+import {QuestionTrackingComponent} from './question/questionTracking/questionTracking.component';
 
 
 const appRoutes = [
@@ -42,6 +43,7 @@ const appRoutes = [
   { path: 'units/:unitId/modules/:moduleId', component: ModuleEditorComponent },
   { path: 'course/:courseId/modules/:moduleId', component: ModuleEditorComponent },
   { path: 'units/:unitId/question', component: QuestionComponent },
+  { path: 'unit/:unitId/question/:questionType/:questionId', component: QuestionTrackingComponent },
   { path: 'answerQuestionDialog', component: AnswerDefinitionDialogComponent},
   { path: 'addQuestionDialog', component: AddQuestionDialogComponent},
   { path: 'units/:unitId/itineraries/:itineraryId/definitionQuestion/:questionId', component: DefinitionQuestionComponent},